NGR-Rail-Lights-and-Operative-outfit-update (#4268)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds Rail lights to the Kali-Andhi class destroyer and the Haboob class
patrol cutter. Additionally makes it so that NGR operatives no longer
spawn wearing helmets and armor vests which they are meant to be
retrieving from the armory.
<!-- Describe The Pull Request. Please be sure every change is
documented, or this can delay review and even discourage maintainers
from merging your PR! -->

## Why It's Good For The Game
Both of these NGR ships lacked on board light sources for their
operatives, which falls out of line with other ships of similar class
and purpose. This provides them with options and looks nice doing it.
Also connected in part to this PR
https://github.com/shiptest-ss13/Shiptest/pull/4266#issue-2877296662.

As for the outfit change, I believe it was an oversight of the created
outfits that the operatives spawned with armor and helmets on,
especially when an exact amount needed to equip them can be found in
each ship's armory. This also brings them in line with the precedent set
by PGF and CMM ships.

Images below:

Rail lights!? Again!?
![Screenshot 2025-02-25
021139](https://github.com/user-attachments/assets/f1d02109-480e-4518-8458-05ca747e4059)

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added lights for the operatives on the haboob and Kali Andhi
fix: fixed operative outfit as to not spawn wearing armor
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>
Co-authored-by: Changelogs <action@github.com>