Lights, Camera, Action - Broadcast Cameras! (#4202)
## About The Pull Request

Adds broadcast cameras, a form of camera capable of one-way audio and
video broadcasting across the sector.
Perfect for hosting a talkshow, performing interviews, livestreaming a
raid or professional journalism on the Frontier.

It comes with a built-in radio, which is capable of broadcasting to
everywhere but can only receive locally.

Uses sprites from CM/SS13's
https://github.com/cmss13-devs/cmss13/pull/5784 PR by GrrrKitten.

Many additional thanks to @zimon9, @Gristlebee, @rye-rice and @Erikafox
for their assistance with development.

## Why It's Good For The Game

Though we have journalists in-code, one of the limitations is that
actual "live broadcasting" has to be abstracted or done via increasingly
hacky or admin-intensive methods. This streamlines that into making a
single device capable of both broadcasting audio and video in a (rather
large) package.

Also, https://youtu.be/f_EouNtFjlU?feature=shared&t=73

## Changelog

:cl:
add: Broadcast Cameras, for sector-wide media journalism!
balance: SecureEye is unrestricted, so that you too can tune into a
livestream!
/:cl: