Reduces minimum timer on C4 from 10 to 0 (#4265)
## About The Pull Request

Reduces the minimum timer that C4 needs to be attached to something to
explode. Do_after to apply it remains unchanged.

## Why It's Good For The Game

Usually imagine C4 as a quick surprise breaching tool compared to slower
ones like Grinding but its gotta sit there for 10 seconds and everyone
can see it and knows

Still gotta spend time to apply it so it shouldn't cause any problems


## Changelog

:cl:
balance: Changed C4 minimum timer from 10 to 0
/:cl:

---------

Co-authored-by: github-actions <action@github.com>