Foodening (Food Desc Rewrite) (#4320)
## About The Pull Request

Rewrites a vast majority of all the food items in the game. Removes a
lot of old cruft, in-jokes, writing errors, racist bits, etc.. Vending
machine snacks are untouched for now, as a revamp of them would require
sprite changes.

## Why It's Good For The Game

A lot of the current food descriptions are either game-y, outdated, or
of the style for the time in the ancient era of the late 2000s. Bringing
it up to the modern era is for the best.

## Changelog

:cl:

del: Removed the borscht misspelling system.
spellcheck: Fixed of typos and errors in multiple food descriptions.
rewrote multiple food descriptions from the ground up.

/:cl: