Basic mob port (#3969)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
port of https://github.com/tgstation/tgstation/pull/60694 
https://github.com/tgstation/tgstation/pull/61455
https://github.com/tgstation/tgstation/pull/70728
https://github.com/tgstation/tgstation/pull/71593
https://github.com/tgstation/tgstation/pull/77143
and other small portions of basic mob code

- [x] port missing subsytem pr so atmos stops running slow
- [ ] figure out that bug that took down the server

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
basic mobs woooo
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
refactor: Adds basic mobs, a replacement for simple mobs that will
hopefully destroy the subtype. This is so far applied to cockroaches,
mice, rats (and their variants)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->