Syndicate Hardsuit Helmet Lights (#4354)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Helmet lights on the Syndicate hardsuit can now be activated
independently of EVA mode.

Renames combat mode to travel mode, removes references to the changing
modes modifying the armor values or speed of the hardsuit since that no
longer exists.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

QoL change to offer more control of when you want your hardsuit light on
or not.

## Changelog

:cl:
add: Syndicate Hardsuit helmet lights can be activated on their own
add: Syndicate Hardsuit no longer references removed mechanics
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->