Adds painkiller gameplay to surgery (#4270)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Surgeries now have a chance (default 50%) to trigger a flinch effect,
which causes extra damage and bleeding
This can be mitigated by medium levels of drunkenness, or neutralized
entirely with a painkiller, including miner's salve
Being asleep in general reduces the chance of triggering a flinch effect
by 20%, and strangling the patient reduces the chance by 40% (you get
one of these, not both)
This does not affect mechanical limbs since they are likely built
significantly better for interior access

## Why It's Good For The Game

Good????? lol

## Changelog

:cl:
add: surgeries to organic limbs that are done without anesthetic or
painkillers now have a chance to cause extra damage
balance: medical cyborgs now have access to morphine in their hypospray
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Theos <theubernyan@gmail.com>